requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-restx==1.1.0
python-dotenv==1.0.0
azure-cosmos==4.5.1
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
pandas==2.0.3
tensorflow==2.13.0
pyspark==3.4.1
python-jose==3.3.0
marshmallow==3.20.1
pytest==7.4.0
black==23.7.0
flake8==6.1.0 